+++ src/components/BookList.tsx
@@ -51,7 +51,6 @@ const hasUrl = (book: BookData): book is BookWithUrl => !!book.url;
 
 export const ListView: React.FC<{
   books: BookData[];
-  breadcrumb?: React.ReactNode;
 }> = ({ books }) => {
   // this hook will refetch the page when we reach the bottom of the screen
   const { listRef, isFetchingPage } = useInfiniteScroll();